revise experiment.py so that target car is an Image object, not just a car_id
