+++ experiment_api/run_experiment.sh
@@ -1,2 +1,1 @@
-# running experiment
 python3 experiment.py -s 5 -c 3 -d 0.3 -e 11 -t 5 -y 2 "/path/to/veri/dataset"
