+++ experiment_api/run_metric.sh
@@ -1,0 +1,1 @@
+python3 metric.py -e 11 -y 2 -r 10000 "/net/dev-vsrv-fs4.b.internal/data1/teams/pelops/veri_unzipped"
