+++ transform_img/install.sh
@@ -1,5 +1,2 @@
 # python 3
 pip3 install python-resize-image docopt
-
-# python 2
-#pip install python-resize-image docopt
